Move UpscalingEngine to .jni
